combine 2 components
